Tratamento de exceção no delete de um User

+++ src/main/java/com/educandoweb/curso/services/UserService.java
@@ -4,36 +4,46 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.curso.entities.User;
 import com.educandoweb.curso.repositories.UserRepository;
+import com.educandoweb.curso.services.exceptions.DataBaseException;
+import com.educandoweb.curso.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository userRepository;
-	
+
 	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (!userRepository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			userRepository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
-		User entity = userRepository.getReferenceById(id); //Pega o Obj. p/ manipular (Sem ser pelo BD)
+		User entity = userRepository.getReferenceById(id); // Pega o Obj. p/ manipular (Sem ser pelo BD)
 		updateData(entity, obj);
 		return userRepository.save(entity); // Salva
 	}
